requirements.txt
fastapiuvicornpydantic